Addded input of data to network.

+++ src/InvestmentPredictor/NeuralNetwork/IWeight.java
@@ -1,12 +1,13 @@
 package InvestmentPredictor.NeuralNetwork;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 public interface IWeight extends Serializable
 {
 	public String GetName();
-	public double GetValue();
-	public double GetWeightValue();
-	public void SetValue(double value);
-	public void SetWeightValue(double weightValue);
+	public BigDecimal GetValue();
+	public BigDecimal GetWeightValue();
+	public void SetValue(BigDecimal value);
+	public void SetWeightValue(BigDecimal weightValue);
 }
