+++ src/InvestmentPredictor/NeuralNetwork/Weight.java
@@ -1,15 +1,17 @@
 package InvestmentPredictor.NeuralNetwork;
 
+import java.math.BigDecimal;
+
 public class Weight implements IWeight
 {
 	// Attributes -----------------------------------------------------
 	private static final long serialVersionUID = -4829604365380322625L;
 	private String _name;
-	private double _value;
-	private double _weightValue;
+	private BigDecimal _value;
+	private BigDecimal _weightValue;
 	
 	// Constructors ---------------------------------------------------
-	public Weight(String name, double value, double weightValue)
+	public Weight(String name, BigDecimal value, BigDecimal weightValue)
 	{
 		this._name = name;
 		this._value = value;
@@ -21,14 +23,14 @@ public class Weight implements IWeight
 	public String GetName() { return this._name; }
 
 	@Override
-	public double GetValue() { return this._value; }
+	public BigDecimal GetValue() { return this._value; }
 	
 	@Override
-	public double GetWeightValue() { return this._weightValue; }
+	public BigDecimal GetWeightValue() { return this._weightValue; }
 	
 	@Override
-	public void SetValue(double value) { this._value = value; }
+	public void SetValue(BigDecimal value) { this._value = value; }
 	
 	@Override
-	public void SetWeightValue(double weightValue) { this._weightValue = weightValue; }
+	public void SetWeightValue(BigDecimal weightValue) { this._weightValue = weightValue; }
 }
